+++ pkg/kube/dialer.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Cockroach Authors
+Copyright 2023 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net"
 	"net/http"
 	"strings"
@@ -160,7 +160,7 @@ func (k *PodDialer) Dial(network, addr string) (net.Conn, error) {
 
 	errorChan := make(chan error)
 	go func() {
-		message, err := ioutil.ReadAll(errStream)
+		message, err := io.ReadAll(errStream)
 		if err != nil {
 			errorChan <- err
 		}