+++ pkg/resource/statefulset_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Cockroach Authors
+Copyright 2023 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ import (
 	"bytes"
 	"flag"
 	"fmt"
-	"io/ioutil"
+	"os"
 	"path/filepath"
 
 	api "github.com/cockroachdb/cockroach-operator/apis/v1alpha1"
@@ -111,7 +111,7 @@ func TestRHImage(t *testing.T) {
 }
 
 func load(t *testing.T, file string) []byte {
-	content, err := ioutil.ReadFile(file)
+	content, err := os.ReadFile(file)
 	if err != nil {
 		t.Fatalf("failed to load yaml file %s: %v", file, err)
 	}