+++ pkg/testutil/builder.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Cockroach Authors
+Copyright 2023 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -48,6 +48,11 @@ func NewBuilder(name string) ClusterBuilder {
 
 func (b ClusterBuilder) Namespaced(namespace string) ClusterBuilder {
 	b.cluster.Namespace = namespace
+	return b
+}
+
+func (b ClusterBuilder) WithAutomountServiceAccountToken(mount bool) ClusterBuilder {
+	b.cluster.Spec.AutomountServiceAccountToken = mount
 	return b
 }